// apps/admin/src/lib/api/generated/models/searchAnalytics.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.31
 */
import type { SearchAnalyticsQuery } from './searchAnalyticsQuery';

export interface SearchAnalytics {
  enabled: boolean;
  flush_interval_seconds: number;
  generated_at: string;
  nohits_queries: SearchAnalyticsQuery[];
  popular_queries: SearchAnalyticsQuery[];
  rules: string[];
}